requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
